Create gar_dolg_pass_gates logic
It's just logic for duty's gates on the garbage, the gates is always closed but not locked. It can be used as a template for your mods, the line 'on_info' is responsible for the conditions under which the gate logic will be executed, in this logic this line is simply commented out.